feat: Complete flashcard application with security enhancements and full CRUD functionality
## 🚀 Major Features Implemented

### 🔐 Login Security Enhancements
- Generic error messages to prevent username enumeration attacks
- IP-based rate limiting (5 attempts, 15-minute blocks)
- Enhanced JWT authentication with secure error handling

### 👤 Modern User Domain Design
- Dual architecture supporting legacy and modern user fields
- Flexible login: email, username, or legacy loginID
- Date of birth instead of age with automatic sync
- Modern user fields: firstName, lastName, username, email
- Database migration with backward compatibility

### 📚 Complete Deck Management
- Full deck CRUD operations with React SPA
- Deck creation with Vietnamese UI and validation
- Deck detail pages with statistics and card management
- Responsive design with loading states

### 🃏 Card Management System
- Complete card creation workflow with React Hook Form + Zod
- Interactive flashcard preview with flip animations
- Real-time deck statistics updates after card operations
- Card list display in deck detail pages
- Fixed bug: Cards now immediately visible after creation

## 🛠️ Technical Improvements

### Backend
- Spring Boot REST API architecture
- Comprehensive validation and error handling
- Security middleware with rate limiting
- Database schema evolution with Flyway
- Real-time deck statistics synchronization

### Frontend
- React 19 + TypeScript + TailwindCSS
- TanStack Query for state management
- Feature-based modular architecture
- Responsive design with accessibility
- Vietnamese localization throughout

### Infrastructure
- Docker containerization for all services
- PostgreSQL database with proper schemas
- CORS configuration for SPA integration
- Comprehensive testing scripts

## 🎯 User Experience
- Seamless navigation between all features
- Real-time updates and optimistic UI patterns
- Error handling with user-friendly Vietnamese messages
- Mobile-responsive design
- Loading states and empty state management

All planned requirements from plan/requirement.md have been successfully implemented and tested.

+++ backend/src/main/java/com/app/flashcard/api/dto/response/UserResponse.java
@@ -2,7 +2,7 @@ package com.app.flashcard.api.dto.response;
 
 public class UserResponse {
     
-    private Integer id;
+    private Long id;
     private String loginId;
     private String name;
     private Integer age;
@@ -10,7 +10,7 @@ public class UserResponse {
     
     public UserResponse() {}
     
-    public UserResponse(Integer id, String loginId, String name, Integer age, String email) {
+    public UserResponse(Long id, String loginId, String name, Integer age, String email) {
         this.id = id;
         this.loginId = loginId;
         this.name = name;
@@ -18,11 +18,11 @@ public class UserResponse {
         this.email = email;
     }
     
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
     
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
